Move parameter and methodcomponent into individual file
