+++ src/main/java/com/amazon/opendistroforelasticsearch/knn/index/KNNMethod.java
@@ -17,11 +17,8 @@ package com.amazon.opendistroforelasticsearch.knn.index;
 
 import org.elasticsearch.common.ValidationException;
 
-import java.util.Collections;
 import java.util.Map;
 import java.util.Set;
-import java.util.function.Function;
-import java.util.stream.Collectors;
 
 /**
  * KNNMethod is used to define the structure of a method supported by a particular k-NN library. It is used to validate
@@ -134,139 +131,5 @@ public class KNNMethod {
      */
     public Map<String, Object> generateExtraParameterMap(KNNMethodContext knnMethodContext) {
         return methodComponent.generateExtraParameterMap(knnMethodContext.getMethodComponent().getParameters());
-    }
-
-    /**
-     * MethodComponent defines the structure of an individual component that can make up an index
-     */
-    public static class MethodComponent {
-
-        private String name;
-        private Map<String, Parameter<?>> parameters;
-
-        /**
-         * Constructor
-         * @param name name of component
-         * @param parameters parameters that the component can take
-         */
-        public MethodComponent(String name, Map<String, Parameter<?>> parameters) {
-            this.name = name;
-            this.parameters = parameters;
-        }
-
-        /**
-         * Get the name of the component
-         *
-         * @return name
-         */
-        public String getName() {
-            return name;
-        }
-
-        /**
-         * Get the parameters for the component
-         *
-         * @return parameters
-         */
-        public Map<String, Parameter<?>> getParameters() {
-            return parameters;
-        }
-
-        /**
-         * Generate extra parameters for this component that are not contained in method string
-         *
-         * @return map of extra parameters
-         */
-        public Map<String, Object> generateExtraParameterMap(Map<String, Object> inParameters) {
-            if (inParameters == null) {
-                return Collections.emptyMap();
-            }
-            return inParameters.entrySet().stream().filter(v -> !parameters.get(v.getKey()).isInMethodString()).collect(
-                    Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
-        }
-
-        /**
-         * Validate that the methodComponentContext is a valid configuration for this methodComponent
-         *
-         * @param methodComponentContext to be validated
-         */
-        public void validate(KNNMethodContext.MethodComponentContext methodComponentContext) {
-            Map<String, Object> providedParameters = methodComponentContext.getParameters();
-            if (providedParameters != null) {
-                for (Map.Entry<String, Object> parameter : providedParameters.entrySet()) {
-                    if (!parameters.containsKey(parameter.getKey())) {
-                        throw new ValidationException();
-                    }
-
-                    parameters.get(parameter.getKey()).validate(parameter.getValue());
-                }
-            }
-        }
-    }
-
-    /**
-     * Parameter that can be set for a method component
-     *
-     * @param <T> Type parameter takes
-     */
-    public static abstract class Parameter<T> {
-
-        private T defaultValue;
-        private boolean inMethodString;
-        protected Function<T, Boolean> validator;
-
-        /**
-         * Constructor
-         *
-         * @param defaultValue of the parameter
-         * @param inMethodString whether the parameter is included in method string
-         */
-        public Parameter(T defaultValue, boolean inMethodString, Function<T, Boolean> validator) {
-            this.defaultValue = defaultValue;
-            this.inMethodString = inMethodString;
-            this.validator = validator;
-        }
-
-        /**
-         * Get default value for parameter
-         *
-         * @return default value of the parameter
-         */
-        public T getDefaultValue() {
-            return defaultValue;
-        }
-
-        /**
-         * Is the parameter included in the method string
-         *
-         * @return true if the parameter should be included in the method string; false otherwise
-         */
-        public boolean isInMethodString() {
-            return inMethodString;
-        }
-
-        /**
-         * Check if the value passed in is valid
-         *
-         * @param value to be checked
-         */
-        public abstract void validate(Object value);
-
-        /**
-         * Integer method parameter
-         */
-        public static class IntegerParameter extends Parameter<Integer> {
-            public IntegerParameter(Integer defaultValue, boolean inMethodString, Function<Integer, Boolean> validator)
-            {
-                super(defaultValue, inMethodString, validator);
-            }
-
-            @Override
-            public void validate(Object value) {
-                if (!(value instanceof Integer) || !validator.apply((Integer) value)) {
-                    throw new ValidationException();
-                }
-            }
-        }
     }
 }
